requirements.txt
flask>=3.0
flask-login>=0.6
libsass>=0.23
gunicorn>=21.0
